% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entsoe-file-api-package.R
\docType{package}
\name{entsoe.file.api-package}
\alias{entsoe.file.api}
\alias{entsoe.file.api-package}
\title{entsoe.file.api: ENTSO-E File API Interface}
\description{
An R package for interfacing with the ENTSO-E file API. This package provides functions to interact with the European Network of Transmission System Operators for Electricity (ENTSO-E) file API with built-in caching and authentication.
}
\keyword{internal}
