% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_entsoe_folders}
\alias{get_entsoe_folders}
\title{Get ENTSO-E Folders}
\usage{
get_entsoe_folders()
}
\value{
A \code{data.table} of the ENTSO-E folders.
}
\description{
This function retrieves the ENTSO-E folders from the 'entsoe_extracts.csv' file within the 'entsoe.file.api' package.
}
\examples{
\dontrun{
get_entsoe_folders()
}
}
