% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{load_entsoe_data}
\alias{load_entsoe_data}
\title{Load ENTSO-E Data}
\usage{
load_entsoe_data(folder, from, to = Sys.time(), checkUpdates = TRUE)
}
\arguments{
\item{folder}{String. The type of ENTSO-E data to load. See \code{\link{get_entsoe_folders}} for available data types.}

\item{from}{POSIXct or character. The start time of the data range.}

\item{to}{POSIXct or character. The end time of the data range. Defaults to the current system time.}

\item{checkUpdates}{Logical. Indicates whether the function should check for data updates. Default is TRUE.}
}
\value{
ENTSO-E data for a given time range as \code{data.table}
}
\description{
This function loads the specified ENTSO-E data for a given time range.
}
\examples{
\dontrun{
load_entsoe_data("EnergyPrices_12.1.D_r3", "2020-01-01", "2020-12-31")
}
}
