% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{set_cache_folder}
\alias{set_cache_folder}
\title{Set Cache Folder}
\usage{
set_cache_folder(folder)
}
\arguments{
\item{folder}{String. The directory to set as the cache folder.}
}
\value{
No return value, called for side effects
}
\description{
This function sets the cache folder for the data files. Calling this function also enables caching.
}
\examples{
\dontrun{
set_cache_folder("your_folder")
}
}
