% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{set_entsoe_credentials}
\alias{set_entsoe_credentials}
\title{Set ENTSO-E API credentials}
\usage{
set_entsoe_credentials(username, password)
}
\arguments{
\item{username}{String. ENTSO-E API username.}

\item{password}{String. ENTSO-E API password.}
}
\value{
No return value, called for side effects
}
\description{
This function sets the username and password for the ENTSO-E API within the e.pkg package.
}
\examples{
\dontrun{
set_entsoe_credentials("your_username", "your_password")
}
}
